Share FakeLookup via tests/render/conftest.py
